compiler warning & layout fix (histogram)

+++ ImageLounge/src/DkMetaData.cpp
@@ -596,7 +596,7 @@ QImage DkMetaDataT::getPreviewImage(int minPreviewWidth) const {
 		for (size_t idx = 0; idx < pList.size(); idx++) {
 			
 			if (pList[idx].width_ > (uint32_t)maxWidth && pList[idx].width_ > (uint32_t)minPreviewWidth) {
-				mIdx = idx;
+				mIdx = (int)idx;
 				maxWidth = pList[idx].width_;
 			}
 		}
